% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couplingsmontecarlo-package.R
\docType{package}
\name{couplingsmontecarlo-package}
\alias{couplingsmontecarlo-package}
\alias{couplingsmontecarlo}
\title{couplingsmontecarlo}
\description{
Couplings and Monte Carlo
}
\details{
...
}
\author{
Pierre E. Jacob <pjacob@fas.harvard.edu>
}
\keyword{package}
